Move `OnAction` to `bindings` package
