+++ tamboui-toolkit/demos/action-handler-demo/src/main/java/dev/tamboui/demo/actionhandler/CounterComponent.java
@@ -8,7 +8,7 @@ import dev.tamboui.style.Color;
 import dev.tamboui.toolkit.component.Component;
 import dev.tamboui.toolkit.element.Element;
 import dev.tamboui.tui.bindings.Actions;
-import dev.tamboui.annotations.OnAction;
+import dev.tamboui.annotations.bindings.OnAction;
 import dev.tamboui.tui.event.Event;
 
 import java.util.function.Consumer;
